fix(demo): track the journal length as a long after InMemoryLedger.size() widened
The restore-aware size() returns long (base events plus tail), so the demo's
refusal check tracks it as one too.

+++ src/main/java/dev/ledgerx/demo/TransferDemo.java
@@ -171,7 +171,7 @@ public final class TransferDemo {
   /** Posts something that must be refused, and proves the ledger did not move. */
   private static void refuse(
       InMemoryLedger ledger, Transaction candidate, RejectionReason expected) {
-    int eventsBefore = ledger.size();
+    long eventsBefore = ledger.size();
     Map<AccountId, Money> balancesBefore = ledger.balances();
     try {
       ledger.post(candidate);
